requirements.txt
appdirs==1.4.3
astroid==1.5.2
backports.functools-lru-cache==1.3
configparser==3.5.0
enum34==1.1.6
gitdb2==2.0.0
GitPython==2.1.3
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
packaging==16.8
pylint==1.7.1
pyparsing==2.2.0
singledispatch==3.4.0.3
six==1.10.0
smmap2==2.0.1
wrapt==1.10.10
